Reestablish kakfa consumer group session on disconnection
